Update messages_zhcn.yml
enhanced some translation
Chinese is a beautiful and complicated language. 
the way last translator translate was to hard? i guess? just not friendly
i did some minor changes, now the language isn't so hard. and i also change the translation of 2fa. 两步验证 is cool. but 双重认证 is more accurate
plz aprove