Fix: Teleporting Players to Spawn on Login with teleportUnAuthedToSpawn=true
Issue: When changing the config option teleportUnAuthedToSpawn=true during gameplay, players were teleported to spawn on their next login. It only happens for the first time after the update.

Patch: AuthMe now checks the player's last logout location from Essentials if it is not found in the AuthMe database. This ensures players are teleported to their last known location instead of the spawn point when the configuration is changed mid-game.